video view/edit/delete with user check

+++ server/src/controllers/video.ts
@@ -10,24 +10,48 @@ import {
 import prisma from "../lib/prisma";
 import { updateVideo } from "../schema/video";
 import { deleteFolder, s3Client } from "../lib/s3Client";
-import { DeleteObjectCommand } from "@aws-sdk/client-s3";
+import { validateToken } from "../middleware";
 
 export async function handleGetVideos(req: Request, res: Response) {
-  const videos = await prisma.video.findMany();
+  const videos = await prisma.video.findMany({
+    where: {
+      visibility: "public",
+    },
+    orderBy: {
+      uploadedAt: "desc",
+    },
+  });
 
-  res.status(200).json({
-    videos: videos.map((video) => ({
+  const { token } = req.cookies;
+
+  const videosResp: any[] = [];
+
+  for (const video of videos) {
+    const key = video.key.split(AWS_BUCKET_NAME + "/")[1];
+    let isAuthor = false;
+    if (token) {
+      const userExists = await validateToken(token);
+      if (userExists && userExists.id === video.userId) {
+        isAuthor = true;
+      }
+    }
+    videosResp.push({
       id: video.id,
       title: video.title,
       description: video.description,
-      url: `https://${AWS_BUCKET_NAME}.s3.${AWS_REGION}.amazonaws.com/${video.key}/index.m3u8`,
+      url: `https://${AWS_BUCKET_NAME}.s3.amazonaws.com/${key}/index.m3u8`,
       likes: video.likes,
       dislikes: video.dislikes,
       views: video.views,
       visibility: video.visibility,
       uploadedAt: video.uploadedAt,
       updatedAt: video.updatedAt,
-    })),
+      isAuthor,
+    });
+  }
+
+  res.status(200).json({
+    videos: videosResp,
   });
 }
 
@@ -44,18 +68,44 @@ export async function handleGetVideo(req: Request, res: Response) {
     return res.status(404).json({ message: "Video not found" });
   }
 
+  const { token } = req.cookies;
+  let isAuthor = false;
+
+  if (token) {
+    const userExists = await validateToken(token);
+    if (userExists && userExists.id === video.userId) {
+      isAuthor = true;
+    }
+  }
+
+  if (video.visibility === "private") {
+    if (!token) {
+      return res.status(401).json({ message: "Unauthorized" });
+    }
+    const userExists = await validateToken(token);
+    if (!userExists) {
+      return res.status(401).json({ message: "Unauthorized" });
+    }
+    if (userExists.id !== video.userId) {
+      return res.status(401).json({ message: "Unauthorized" });
+    }
+  }
+
+  const key = video.key.split(AWS_BUCKET_NAME + "/")[1];
+  console.log("Key", key);
   res.status(200).json({
     video: {
       id: video.id,
       title: video.title,
       description: video.description,
-      url: `https://${AWS_BUCKET_NAME}.s3.${AWS_REGION}.amazonaws.com/${video.key}/index.m3u8`,
+      url: `https://${AWS_BUCKET_NAME}.s3.amazonaws.com/${key}/index.m3u8`,
       likes: video.likes,
       dislikes: video.dislikes,
       views: video.views,
       visibility: video.visibility,
       uploadedAt: video.uploadedAt,
       updatedAt: video.updatedAt,
+      isAuthor,
     },
   });
 }
@@ -107,6 +157,7 @@ export async function handleUploadVideo(req: Request, res: Response) {
 
 export async function handleUpdateVideo(req: Request, res: Response) {
   const { id } = req.params;
+  const { user } = res.locals;
   const params = updateVideo.safeParse(req.body);
 
   if (!params.success) {
@@ -121,6 +172,10 @@ export async function handleUpdateVideo(req: Request, res: Response) {
 
   if (!video) {
     return res.status(404).json({ message: "Video not found" });
+  }
+
+  if (video.userId !== user.id) {
+    return res.status(401).json({ message: "Unauthorized" });
   }
 
   const { title, description, visibility } = params.data;
@@ -144,6 +199,7 @@ export async function handleUpdateVideo(req: Request, res: Response) {
 
 export async function handleDeleteVideo(req: Request, res: Response) {
   const { id } = req.params;
+  const { user } = res.locals;
 
   const video = await prisma.video.findUnique({
     where: {
@@ -155,7 +211,11 @@ export async function handleDeleteVideo(req: Request, res: Response) {
     return res.status(404).json({ message: "Video not found" });
   }
 
-  const deleteRes= await deleteFolder(video.key);
+  if (video.userId !== user.id) {
+    return res.status(401).json({ message: "Unauthorized" });
+  }
+
+  const deleteRes = await deleteFolder(video.key);
 
   if (!deleteRes) {
     return res.status(500).json({ message: "Failed to delete video" });
